* sysdeps/unix/sysv/linux/mips/mips64/n32/sysdep.h 	(INTERNAL_SYSCALL): Update internal_syscall##nr invocation. 	(INTERNAL_SYSCALL_NCS): New. 	(internal_syscall0, internal_syscall1, internal_syscall2, 	internal_syscall3, internal_syscall4, internal_syscall5, 	internal_syscall6): Take ncs_init, cs_init, and input arguments. 	Use them.  Correct types for registers. 	* sysdeps/unix/sysv/linux/mips/mips64/n64/sysdep.h 	(INTERNAL_SYSCALL): Update internal_syscall##nr invocation. 	(INTERNAL_SYSCALL_NCS): New. 	(internal_syscall0, internal_syscall1, internal_syscall2, 	internal_syscall3, internal_syscall4, internal_syscall5, 	internal_syscall6): Take ncs_init, cs_init, and input arguments. 	Use them. 	* sysdeps/unix/sysv/linux/mips/mips64/syscalls.list: Remove 	recvfrom and sendto.  Mark lseek, msgrcv, and msgsnd as cancellation 	points. 	* sysdeps/mips/dl-machine.h (elf_machine_rel): Remove unused "value". 	Use Elf(Addr) for TLS relocation targets. 	* sysdeps/unix/sysv/linux/mips/mips64/Makefile: New file. 	* sysdeps/unix/sysv/linux/mips/nptl/lowlevellock.h (lll_futex_wait, 	lll_futex_timed_wait, lll_futex_wake, lll_futex_requeue): Cast 	futexp to long for n64. 	* sysdeps/unix/sysv/linux/mips/mips64/nptl/sysdep-cancel.h: New file.

+++ sysdeps/unix/sysv/linux/mips/nptl/lowlevellock.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2003, 2004, 2005 Free Software Foundation, Inc.
+/* Copyright (C) 2003, 2004, 2005, 2006 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -39,7 +39,7 @@
     INTERNAL_SYSCALL_DECL (__err);					      \
     long int __ret;							      \
     __ret = INTERNAL_SYSCALL (futex, __err, 4,				      \
-			      (futexp), FUTEX_WAIT, (val), 0);		      \
+			      (long) (futexp), FUTEX_WAIT, (val), 0);	      \
     INTERNAL_SYSCALL_ERROR_P (__ret, __err) ? -__ret : __ret;		      \
   })
 
@@ -48,7 +48,7 @@
     INTERNAL_SYSCALL_DECL (__err);					      \
     long int __ret;							      \
     __ret = INTERNAL_SYSCALL (futex, __err, 4,				      \
-			      (futexp), FUTEX_WAIT, (val), (timespec));	      \
+			      (long) (futexp), FUTEX_WAIT, (val), (timespec));\
     INTERNAL_SYSCALL_ERROR_P (__ret, __err) ? -__ret : __ret;		      \
   })
 
@@ -57,7 +57,7 @@
     INTERNAL_SYSCALL_DECL (__err);					      \
     long int __ret;							      \
     __ret = INTERNAL_SYSCALL (futex, __err, 4,				      \
-			      (futexp), FUTEX_WAKE, (nr), 0);		      \
+			      (long) (futexp), FUTEX_WAKE, (nr), 0);	      \
     INTERNAL_SYSCALL_ERROR_P (__ret, __err) ? -__ret : __ret;		      \
   })
 
@@ -67,7 +67,7 @@
     INTERNAL_SYSCALL_DECL (__err);					      \
     long int __ret;							      \
     __ret = INTERNAL_SYSCALL (futex, __err, 6,				      \
-			      (futexp), FUTEX_CMP_REQUEUE, (nr_wake),	      \
+			      (long) (futexp), FUTEX_CMP_REQUEUE, (nr_wake),  \
 			      (nr_move), (mutex), (val));		      \
     INTERNAL_SYSCALL_ERROR_P (__ret, __err);				      \
   })
